[bcr-wallet-core] add reclaim_funds API
to reclaim all funds that have been marked as Pending or PendingSpent.
The wallet cross-checks with the mint the current status of the proofs,
if they are still unspent, we claim them back and scoop the funds.

This commit also adds some minor refactoring of other APIs, with some
rename, APIs aggregation, and js example refactor